// src/app/modules/system/interfaces/process.ts
export interface Process {
  createDate?: any;
  createdTime?: Date;
  data?: any;
  description?: string;
  duration?: any;
  endDate?: any;
  endTime?: Date;
  filename?: string;
  files?: any;
  guid?: string;
  log?: string;
  message?: string;
  name?: string;
  notify?: string;
  percent?: number;
  pid?: string;
  prettyName?: any;
  id?: number;
  processState?: any;
  startDate?: any;
  startTime?: Date;
  state?: string;
  stats?: string;
  type?: string;
}